+++ src/api/email_verification.py
@@ -67,11 +67,6 @@ def update_registration_status():
           schema:
             type: string
           required: true
-        - name: redirect_uri
-          in: query
-          schema:
-            type: string
-            format: uri
     responses:
         200:
             description: OK
@@ -81,8 +76,7 @@ def update_registration_status():
             description: Unexpected error.
     """
     email_token = request.args.get("token", "")
-    redirect_uri = request.args.get("redirect_uri",
-                                    app.config.get("FRONTEND_URL"))
+    redirect_uri = app.config.get("FRONTEND_URL")
 
     if not email_token:
         raise BadRequest("No email token was provided")