+++ CHANGELOG.md
@@ -16,7 +16,17 @@ Types of Changes:
 
 ## [Unreleased]
 
+### Deprecated
+
+- The use of `multipart/form-data` on the POST `/api/hackers/` endpoint.
+
+### Added
+
+-   POST `/api/hackers/resume/` endpoint for uploading a hacker's resume before the hacker is actually created in the DB. Unattached resumes are deleted after 24 hours.
+-   Added the ability to upload the hacker document as a `application/json` while providing the optional `resume_id` value to attach a resume to a hacker document which is returned from the `/api/hackers/resume/` endpoint.
+
 ### Fixed
+
 -   Fixed get hacker resume field returning a 500 error.
 
 ### Security